+++ lascli/parser/__init__.py
@@ -1,6 +1,7 @@
 from .app_clients import create_app_clients_parser
 from .assets import create_assets_parser
 from .datasets import create_datasets_parser
+from .deployment_environments import create_deployment_environments_parser
 from .documents import create_documents_parser
 from .logs import create_logs_parser
 from .models import create_models_parser