+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Version 9.5.0 - 2023-04-18
+
+- Added `deployment-environments get`
+- Added `deployment-environments list`
+- Added `--postprocess-config` to `models create`
+- Updated `--preprocess-config` in `models create` to also accept inline JSON
+- Added `--postprocess-config` to `models update`
+- Updated `--preprocess-config` in `models update` to also accept inline JSON
+- Updated `--postprocess-config` in `predictions create` to also accept path to a JSON file
+
 ## Version 9.4.0 - 2023-02-23
 
 - Added `--rotation` to `predictions create`